+++ apps/plugins/sort.c
@@ -59,7 +59,7 @@ PLUGIN_HEADER
 
 static struct plugin_api* rb;
 
-int buf_size;
+ssize_t buf_size;
 static char *filename;
 static int num_entries;
 static char **pointers;
